Some additional structures & removed useless files
At the beginning I thought to create a very basic internal web server for development only in addition to fcgi. This idea of this little server was to make the web development using this web framework a bit easier but it would be a big a mess to allow the developer to switch between fcgi and the internal server because they would be completely different!
But this idea can still be possible at the end of the development during the refactoring.
Let the surprise come.